# website/static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Last-mile network emulator control plane for Docker sidecars.\nOptional Bearer auth when POTATONETWORK_API_TOKEN is set.",
        "title": "PotatoNetwork API",
        "contact": {
            "name": "PotatoNetwork",
            "url": "https://github.com/kriakiku/potato-network"
        },
        "license": {
            "name": "See repository LICENSE",
            "url": "https://github.com/kriakiku/potato-network"
        },
        "version": "1.0"
    },
    "host": "localhost:7783",
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "description": "302 to GitHub Pages OpenAPI section",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Redirect to docs",
                "responses": {
                    "302": {
                        "description": "Location: docs URL",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/baseline": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "baseline"
                ],
                "summary": "Get baseline",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.BaselineResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "baseline"
                ],
                "summary": "Set baseline",
                "parameters": [
                    {
                        "description": "Host RTT map",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.BaselinePutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.BaselineResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/baseline/probe": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "baseline"
                ],
                "summary": "Probe baseline",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.BaselineResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/ca.pem": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/x-pem-file"
                ],
                "tags": [
                    "ca"
                ],
                "summary": "Download CA certificate",
                "responses": {
                    "200": {
                        "description": "PEM certificate",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/catalog": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catalog"
                ],
                "summary": "Get catalog",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/catalog.Catalog"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/catalog/refresh": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catalog"
                ],
                "summary": "Refresh catalog",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.CatalogRefreshResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/health": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.HealthResponse"
                        }
                    }
                }
            }
        },
        "/v1/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Get profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/profiles.Profile"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Set profile",
                "parameters": [
                    {
                        "description": "Country+tier or passthrough",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ProfilePutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/profiles.Profile"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/rules": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rules"
                ],
                "summary": "Rules status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.RulesStatus"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.BaselinePutRequest": {
            "type": "object",
            "properties": {
                "hostRtt": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                }
            }
        },
        "api.BaselineResponse": {
            "type": "object",
            "properties": {
                "hostRtt": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "probedAt": {
                    "type": "string",
                    "example": "2026-01-01T00:00:00Z"
                }
            }
        },
        "api.CatalogRefreshResponse": {
            "type": "object",
            "properties": {
                "generatedAt": {
                    "type": "string",
                    "example": "2026-01-01T00:00:00Z"
                },
                "ok": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "unauthorized"
                }
            }
        },
        "api.HealthResponse": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean",
                    "example": true
                },
                "product": {
                    "type": "string",
                    "example": "PotatoNetwork"
                },
                "profile": {
                    "type": "string",
                    "example": "passthrough"
                },
                "shape": {
                    "type": "string",
                    "example": "passthrough"
                }
            }
        },
        "api.ProfilePutRequest": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string",
                    "example": "BD"
                },
                "passthrough": {
                    "type": "boolean",
                    "example": false
                },
                "tier": {
                    "type": "string",
                    "example": "typical"
                }
            }
        },
        "api.RulesStatus": {
            "type": "object",
            "properties": {
                "compile": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean",
                    "example": true
                },
                "path": {
                    "type": "string",
                    "example": "/data/rules.expr"
                },
                "source": {
                    "type": "string"
                }
            }
        },
        "catalog.Catalog": {
            "type": "object",
            "properties": {
                "countries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/catalog.Country"
                    }
                },
                "destinations": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/catalog.Destination"
                    }
                },
                "generatedAt": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                }
            }
        },
        "catalog.Country": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nearestAws": {
                    "type": "string"
                },
                "rttSource": {
                    "type": "string"
                },
                "tiers": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/catalog.Tier"
                    }
                }
            }
        },
        "catalog.Destination": {
            "type": "object",
            "properties": {
                "awsRegion": {
                    "type": "string"
                },
                "calibrate": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "target": {
                    "type": "string"
                }
            }
        },
        "catalog.Tier": {
            "type": "object",
            "properties": {
                "downloadMbps": {
                    "type": "number"
                },
                "lossPercent": {
                    "type": "number"
                },
                "rttToDest": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "uploadMbps": {
                    "type": "number"
                }
            }
        },
        "profiles.Profile": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "delayMs": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "downloadMbps": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "lossPercent": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "passthrough": {
                    "type": "boolean"
                },
                "tier": {
                    "type": "string"
                },
                "uploadMbps": {
                    "type": "number"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Optional. Format: Bearer followed by the API token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}